requirements.txt
alembic==1.9.2
appdirs==1.4.4
astroid==2.5.1
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
distlib==0.3.1
et-xmlfile==1.1.0
Faker==37.3.0
filelock==3.0.12
Flask==1.1.2
Flask-Excel==0.0.7
Flask-HTTPAuth==4.7.0
Flask-Login==0.5.0
Flask-Migrate==4.0.2
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==2.0.1
idna==2.10
importlib-metadata==6.0.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
lml==0.1.0
Mako==1.2.4
MarkupSafe==1.1.1
marshmallow==3.0.0
mccabe==0.6.1
mysql-connector-python==9.3.0
mysqlclient==2.2.7
numpy==1.26.3
openpyxl==3.1.2
packaging==23.0
pandas==2.1.4
pycodestyle==2.7.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
PyJWT==1.7.1
pylint==2.7.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
texttable==1.7.0
toml==0.10.2
twilio==6.56.0
tzdata==2023.4
urllib3==1.26.4
virtualenv==20.4.2
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
XlsxWriter==3.1.9
zipp==3.11.0
